feat: sync wishlist after pruning
