+++ massa-consensus-worker/src/state/mod.rs
@@ -10,6 +10,7 @@ use massa_consensus_exports::{
     ConsensusChannels, ConsensusConfig,
 };
 use massa_execution_exports::ExecutionBlockMetadata;
+use massa_logging::massa_trace;
 use massa_metrics::MassaMetrics;
 use massa_models::{
     active_block::ActiveBlock,
@@ -528,6 +529,52 @@ impl ConsensusState {
         }
 
         Ok(wishlist)
+    }
+
+    /// Recompute the block wishlist from the currently retained waiting blocks
+    /// and publish the delta (new and removed entries) to protocol.
+    ///
+    /// Invariant: the published wishlist is always derived exclusively from
+    /// currently retained, reachable waiting blocks. This must be called after
+    /// any operation that removes or invalidates waiting-for-dependency blocks
+    /// (e.g. `block_db_changed`, `prune_waiting_for_dependencies`) so that
+    /// protocol does not keep requesting blocks that consensus has already
+    /// discarded as unreachable.
+    ///
+    /// # Returns
+    /// Error if the wishlist could not be computed or sent.
+    pub fn sync_wishlist(&mut self) -> Result<(), ConsensusError> {
+        // compute the block wishlist from currently retained waiting blocks
+        let new_wishlist = self.get_block_wishlist()?;
+        let new_blocks: PreHashMap<BlockId, Option<SecuredHeader>> = new_wishlist
+            .iter()
+            .filter_map(|(id, header)| {
+                if !self.wishlist.contains_key(id) {
+                    Some((*id, header.clone()))
+                } else {
+                    None
+                }
+            })
+            .collect();
+        let remove_blocks: PreHashSet<BlockId> = self
+            .wishlist
+            .iter()
+            .filter_map(|(id, _)| {
+                if !new_wishlist.contains_key(id) {
+                    Some(*id)
+                } else {
+                    None
+                }
+            })
+            .collect();
+        if !new_blocks.is_empty() || !remove_blocks.is_empty() {
+            massa_trace!("consensus.consensus_worker.sync_wishlist.send_wishlist_delta", { "new": new_wishlist, "remove": remove_blocks });
+            self.channels
+                .protocol_controller
+                .send_wishlist_delta(new_blocks, remove_blocks)?;
+            self.wishlist = new_wishlist;
+        }
+        Ok(())
     }
 
     /// Gets a block and all its descendants
